started splitting and grouping variable and dataset operations
