+++ python/scipp.cpp
@@ -6,18 +6,21 @@
 
 namespace py = pybind11;
 
+void init_counts(py::module &);
 void init_dataset(py::module &);
+void init_dataset_free_functions(py::module &);
 void init_detail(py::module &);
 void init_dtype(py::module &);
 void init_eigen(py::module &);
-void init_counts(py::module &);
+void init_element_array_view(py::module &);
+void init_event_list(py::module &);
 void init_groupby(py::module &);
 void init_neutron(py::module &);
 void init_operations(py::module &);
-void init_event_list(py::module &);
+void init_reduction(py::module &);
 void init_units_neutron(py::module &);
 void init_variable(py::module &);
-void init_element_array_view(py::module &);
+void init_variable_free_functions(py::module &);
 
 void init_core(py::module &m) {
   auto core = m.def_submodule("core");
@@ -29,8 +32,11 @@ void init_core(py::module &m) {
   init_dataset(core);
   init_groupby(core);
   init_operations(core);
+  init_reduction(core);
   init_event_list(core);
   init_element_array_view(core);
+  init_variable_free_functions(core);
+  init_dataset_free_functions(core);
 }
 
 PYBIND11_MODULE(_scipp, m) {
