+++ python/tests/test_dtype.py
@@ -14,7 +14,7 @@ def test_dtype():
 
 
 @pytest.mark.skip(reason="Unfortunately the scipp dtype is currently not \
-        compatible with the numpy dtype. Scippy supports types such as \
+        compatible with the numpy dtype. Scipp supports types such as \
         strings which numpy cannot handle, so we cannot simply use \
         numpy.dtype.")
 def test_numpy_comparison():
