+++ python/trigonometry.cpp
@@ -1,0 +1,35 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// Copyright (c) 2020 Scipp contributors (https://github.com/scipp)
+/// @file
+/// @author Simon Heybrock
+#include "pybind11.h"
+
+#include "scipp/dataset/dataset.h"
+#include "scipp/variable/operations.h"
+
+using namespace scipp;
+using namespace scipp::variable;
+using namespace scipp::dataset;
+
+namespace py = pybind11;
+
+template <class T> void bind_flatten(py::module &m) {
+  using ConstView = const typename T::const_view_type &;
+  m.def("flatten", py::overload_cast<ConstView, const Dim>(&flatten),
+        py::arg("x"), py::arg("dim"), py::call_guard<py::gil_scoped_release>(),
+        R"(
+        Flatten the specified dimension into event lists, equivalent to summing dense data.
+
+        :param x: Variable, DataArray, or Dataset to flatten.
+        :param dim: Dimension over which to flatten.
+        :raises: If the dimension does not exist, or if x does not contain event lists
+        :seealso: :py:class:`scipp.sum`
+        :return: New variable, data array, or dataset containing the flattened data.
+        :rtype: Variable, DataArray, or Dataset)");
+}
+
+void init_operations(py::module &m) {
+  bind_flatten<Variable>(m);
+  bind_flatten<DataArray>(m);
+  bind_flatten<Dataset>(m);
+}
